refacto offerers cleant
