+++ src/components/pages/OfferersPage.js
@@ -10,9 +10,9 @@ import { connect } from 'react-redux'
 import { compose } from 'redux'
 import { NavLink } from 'react-router-dom'
 
+import OffererItem from './offerers/OffererItem'
 import HeroSection from '../layout/HeroSection'
 import Main from '../layout/Main'
-import OffererItem from '../items/OffererItem'
 import offerersSelector from '../../selectors/offerers'
 import { offererNormalizer } from '../../utils/normalizers'
 import { mapApiToWindow, windowToApiQuery } from '../../utils/pagination'
